// src/stage/types/tile.ts
export enum Tile {
  SKY = "SKY",
  GROUND_TOP_LEFT = "GROUND_TOP_LEFT",
  GROUND_TOP_RIGHT = "GROUND_TOP_RIGHT",
  GROUND_TOP_MIDDLE = "GROUND_TOP_MIDDLE",
  PIPE_TOP_LEFT = "PIPE_TOP_LEFT",
  PIPE_TOP_RIGHT = "PIPE_TOP_RIGHT",
  PIPE_BOTTOM_LEFT = "PIPE_BOTTOM_LEFT",
  PIPE_BOTTOM_RIGHT = "PIPE_BOTTOM_RIGHT",
  BRICK1 = "BRICK1",
}